refactor(arch): fix import paths after domain module rename (#209)
Update all broken relative imports following directory renames:
- ext/ -> vscode/, actions/ -> journal/ (name changes)
- provider/commands/ -> commands/ (depth 3->2, all ../../ -> ../)
- provider/features/sync-* -> features/sync/
- provider/features/scan-entries -> features/entries/
- provider/features/match-input -> journal/
- provider/codeactions/ -> ui/codeactions/
- util/paths.ts sibling imports -> ../util/ cross-dir imports

#209
