+++ src/test/suite/commands-copy-task.test.ts
@@ -1,7 +1,7 @@
 import * as assert from 'assert';
 import * as vscode from 'vscode';
 import * as J from '../..';
-import { ShiftTarget, CopyTaskCommand } from '../../provider/commands/copy-task';
+import { ShiftTarget, CopyTaskCommand } from '../../commands/copy-task';
 import { createMockCtrl } from './command-test-helpers';
 
 suite('Command suites - copy task command', () => {
