+++ src/test/suite/commands-entry.test.ts
@@ -3,10 +3,10 @@ import * as os from 'os';
 import * as path from 'path';
 import * as vscode from 'vscode';
 import * as J from '../..';
-import { ShowEntryForInputCommand } from '../../provider/commands/show-entry-for-input';
-import { ShowEntryForTodayCommand } from '../../provider/commands/show-entry-for-today';
-import { ShowEntryForTomorrowCommand } from '../../provider/commands/show-entry-for-tomorrow';
-import { ShowEntryForYesterdayCommand } from '../../provider/commands/show-entry-for-yesterday';
+import { ShowEntryForInputCommand } from '../../commands/show-entry-for-input';
+import { ShowEntryForTodayCommand } from '../../commands/show-entry-for-today';
+import { ShowEntryForTomorrowCommand } from '../../commands/show-entry-for-tomorrow';
+import { ShowEntryForYesterdayCommand } from '../../commands/show-entry-for-yesterday';
 import { createMockCtrl, tick } from './command-test-helpers';
 import { TestLogger } from '../test-logger';
 import { fileExists } from '../../util/fs-exists';
